added a category validation

+++ src/main/java/com/example/productservice/category/CategoryService.java
@@ -3,6 +3,7 @@ package com.example.productservice.category;
 
 import com.example.productservice.category.exception.CategoryAlreadyExistsException;
 import com.example.productservice.category.exception.CategoryNotFoundException;
+import com.example.productservice.product.ProductRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +15,7 @@ import java.util.stream.Collectors;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final ProductRepository productRepository;
 
     public void createNewCategory(String categoryName){
         if (categoryRepository.existsByCategoryName(categoryName)){
@@ -35,7 +37,11 @@ public class CategoryService {
     public void deleteCategory(Long categoryId){
         Category categoryToDelete = categoryRepository.findById(categoryId)
                 .orElseThrow(() -> new CategoryNotFoundException(categoryId));
-//         TODO: 7/12/2023 check if there is any item in this category
+
+        if (productRepository.existsProductByCategory_Id(categoryId)) {
+            throw new IllegalArgumentException("Cannot delete category with ID " + categoryId + ". " +
+                    "You have to delete products associated with this category first");
+        }
         categoryRepository.delete(categoryToDelete);
     }
